+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added Collection extensions for setting and adding ranges of items
  - Added a method for getting the total number of seconds in a date
  - Added extensions to raise events with common event args using the data directly
+ - Added property to IcdEnvironment to determine whether SSL communication is enabled
 
 ### Changed
  - Repeater changed to use configured callbacks instead of a dumb event